I have a user (David Trudgian) who experienced some trouble using the local sudo installer behind a proxy.
He got it to work by adding the following lines to /etc/sudoers

    Defaults env_keep +="http_proxy"
    Defaults env_keep +="https_proxy"

I think this changeset may have the same effect (I'll see if he is willing to test it), but not require modifications to /etc/sudoers. If this is overly specific, feel free to pull this out and I will find a place to document modifing /etc/sudoers instead.