Add function to generalize biotype lookup.
Look in either the source column or in the gene_biotype info
field for protein_coding, if it is in either one, then use that
as the biotype column.

Also fix for replacing rnaseq directory with a link
if a directory already exists

Thanks to @zeneofa for the bug report.