Add bioconductor-snpchip to ngs_pipeline_minimal.
TitanCNA requires bioconductor-snpchip for their wrapper script, but the
authors don't want to add this as a dependency, since it isn't strictly
necessary for TitanCNA to run. Rather than continually fixing the
bioconda formula, we'll just make sure bcbio installs it here.

Fixes bcbio/bcbio-nextgen#2874. Sorry @lbeltrame for taking so long to come to this
solution.