Add `buildversion` option to prepare_tx_gff.
This lets an arbitrary string get passed on as the build version.
This is helpful when making a custom genome for identifying which
GTF file was used to make the genome. The string is freeform, but
it is recommended to provide source_version type labelling such
as Ensembl_95 or Flybase_21, etc.